+++ object.h
@@ -9,18 +9,18 @@ class Object
 private:
 	int classID;
     std::string className;
-    std::vector<float> features;
+    std::vector<double> features;
 
 
 public:
 
-    Object(const std::string &className, const std::vector<float> &features) :classID(-1), className(className), features(features)
+    Object(const std::string &className, const std::vector<double> &features) :classID(-1), className(className), features(features)
     {
     }
 
     std::string getClassName() const;
     size_t getFeaturesNumber() const;
-    const std::vector<float> &getFeatures() const;
+    const std::vector<double> &getFeatures() const;
 };
